+++ CSM Foundation Database/CHANGELOG.md
@@ -1,5 +1,19 @@
 # CSM Foundation Database CHANGELOG
 
+## [1.1.0] - 06.08-2025
+
+### Added
+
+- A new [ActivationReference] attribute was created to handle database contexts DbSet usage with interfaces setting a default reference as the activation at the database activation process.
+
+#### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 1.2.1            | 1.3.0           |
+| Microsoft.EntityFrameworkCore           | 9.0.8            | 9.0.8           |
+| Microsoft.EntityFrameworkCore.SqlServer | 9.0.8            | 9.0.8           |
+
 ## [1.0.0] - 06.08-2025
 
 ### Added